ir: Print comment for unnamed basic blocks instead of the label name.
There seems to be an issue with how LLVM handles local IDs, where
labels with explicit local ID names are not accounted for when
incrementing the local ID counter.

A label with the explicit name "0" is treated as a named value
and the local ID 0 may be assigned to subsequent basic blocks or
other values. In contrast, a local variable with the explicit
name "%0" is treated as an unnamed value, and assigned the local
ID 0.

Updates issue #72.

are not treated


Former-commit-id: abec15b3b2198a7f24d103c98098f7ec1c49ac9b