ir: Merge instruction package with ir package to resolve long standing issues.
Prior to this merge, the return type of inst.Parent() had to be value.Value
instead of the more intuitive *BasicBlock type, since referring to
*ir.BasicBlock from the instruction package would create an import cycle.

The constant binary operations will remain in package constant to avoid
confusion and name collisions. Since they won't be referring to basic blocks,
not being able to import package ir is not estimated to become an issue.