Copy parser.go and tree.go without modification from inspirer/textmapper@cbc923c1d47fa589f0a6a1e9372ec229dab5ea6d.
@inspirer I hope it's ok that I copy these files?

I started to look at using arrow notation for
generating the AST. However, from what I understood,
there is no automatically generated code that produces
the tree. And the listener receives NodeType rather
than Node.

I know you are still writing the documentation for the
more advanced features of Textmapper. I wanted to try
using Textmapper and looked for examples on generating
the AST, and the only example I could find was in
textmapper/tm-go/parser.

So, to get up to speed more quickly, and evaluate how
to use this approach, I'll simply base the parser from
the code you wrote.

As a notice for the Git history. The code added in this
commit is licensed Apache 2.0 License.


Former-commit-id: da2619a6a3889206b18885bbde82c7adbe2dd5e5