enc: copy enc without modification from llir/l@50f33ef0f5c32b87ef69df42326c5d6f69c93b04 [formerly 31e24a3e7a2b61d2e981af92c6446db3e6f6ef04].
Note, this copy will be removed, once the repos are merged.

Alternatively. The parser repo will be kept distinct from the
IR repo, mainly to keep the size of the repo down. As the
generated parser and lexer are quite large, it may be wise
to keep the autogenerated code in a contained repo; e.g.
llir/parser or llir/ll or llir/asm. This way, we can force
push to that repo whenever the parser is regenerated, and keep
the size of the main llir/llvm repo manageable.


Former-commit-id: aadbc6bfe7f95a1a2e5ba9dd8595d0f20351a385