grammar: use Type Value instead of TypeValue in Invokee
This is so that we may retrieve the type (before the body)
when creating local variable skeletons. Also, the type
may either be the return type of the invokee, or the function
signature. In the former case, using TypeValue is confusing,
as the type is not actually the type of the value.

	lalr: 0.2s, text: 0.556s, parser: 2164 states, 186KB


Former-commit-id: 866c96b00cfc2ed5c70a0863e26a450f22bd1be8