ir: always return something useful from GlobalIdent.Name
Prior to this commit, calling f.Name() or global.Name()
on an *ir.Function or *ir.Global, the returned string was
empty if the function or global was unnamed, i.e. @1.

To make the Name method more useful, and limit users of
the IR representation from having to re-implement logic
that strips the "@" prefix from Ident() just to get the
string used as name for a global value, implement this
once in GlobalIdent.Name.

I personally wish that the LLVM IR representation would
not allow numeric names to avoid confusion with unnamed
IDs, but since it is supported, we should support it too.