ir: add MDAttachments method to provide access to metadata
Users may define an interface to gain access without
having to do type assertion to the concrete underlying
types.

	type valueWithMetadata interface {
		MDAttachments() []*metadata.Attachment
	}

	if i, ok := inst.(valueWithMetadata); ok {
		mds := i.MDAttachments()
		// use mds.
	}

In the future, perhaps we may want to change the Instruction
and Terminator interfaces to include the MDAttachments method.
This way, users could directly gain access to metadata without
the need of type assertions.