ir: simplify metadata API (part 1)
Update NamedMetadataDefs to be a map from
metadata name to named metadata definition
instead of a slice.

Add metadata IDs to metadata definitions (i.e.
metadata tuple and specialized metadata nodes).

This is in preparation for a simplification of
the specialized metadata API, to make it possible
for users to access DWARF debug info fields
without having to use type assertions as often.