ir/metadata: use *DIFile instead of Field in DICompileUnit
This commit provides a template for the simplifications
planned for the metadata API.

In particular, the main focus is to make life easier for
users of the metadata API, so they should not have to do
as many type assertions to reach the "real" data of Debug
Info specialized metadata nodes.

The fields of the remaining specialized metadata nodes
will be simplified in the same fashion in follow up
commits.