ir: allow callee of function type in NewCall
Prior to this commit, the callee had to be of pointer
to function type (as is the type of functions declared
in the global scope of the module).

However, to allow inline asm values to have function
type instead of pointer to function type, the type
checking of NewCall is relaxed.

Fixes #33.