Update x/tools/cmd/stringer
As of
https://github.com/golang/tools/commit/63e6ed9258fa6cbc90aab9b1eef3e0866e89b874
stringer includes a compile time check that the constants have not changed since the last run.

I just did a rebuild of `go generate` and ended up with this large unwanted diff, so I think it is probably best to stay up-to-date with this tool.

As far as I see, this has little impact on anyone other than llvm/llir devs, who just need to make sure they keep cmd/stringer uptodate.