add initial HACKING.md documentation
This document aims to introduce the various components of
the llir/llvm project and give the curious users better
insight into what goes on under the hood.

We should probably create an associated DESIGN.md document
to provide high-level motivation for various design decisions
and API choises. It could also be good to start defining the
projects goals and non-goals.