internal/natsort: refine natural sorting order
Prior to this commit, the array `a` as defined below:

	a := []string{
		"foo20",
		"foo.bar",
		"foo2",
		"foo.10",
		"foo.1",
		"foo.20",
		"foo.11",
		"foo1",
		"foobar",
		"foo21",
		"foo10",
		"foo11",
		"foo.21",
		"foo.2",
	}

would be sorted as follows:

	"foo1"
	"foo2"
	"foo10"
	"foo11"
	"foo20"
	"foo21"
	"foo.1"
	"foo.2"
	"foo.10"
	"foo.11"
	"foo.20"
	"foo.21"
	"foo.bar"
	"foobar"

That is, digits would be considered "less" than
other characters.

After this commit, the same slice would be sorted as
follows:

	"foo.1"
	"foo.2"
	"foo.10"
	"foo.11"
	"foo.20"
	"foo.21"
	"foo.bar"
	"foo1"
	"foo2"
	"foo10"
	"foo11"
	"foo20"
	"foo21"
	"foobar"

This conforms to the sorting behaviour of many
File Managers and may thus be considered expected
behaviour for users.

Note, this is the first commit to internal/natsort
that diverges in functionality from the sortorder
package of @fvbommel. As such, natsort should now
be considered a fork and not just a vendored mirror.

As stated in the README.md document of the
internal/natsort directory, the original code written
by Frits van Bommel is licensed under an MIT license.
Any subsequent changes made to the code base after
the fork are hereby released into the public domain.

This commit also extends the test cases.