asm,ir: fix handling of index vector length in gep
The issue was twofold, firstly, the vector length was only checked for
index when indexing into struct types, it should always be checked
(including for the 0:th index!). The second issue was that the conversion
from constant.Constant, value.Value and ast.TypeValue omited the index
vector length for indices not considered suited for use to index into
structure types. However, as already concluded, the vector length is
relevent for indexing into other fields and just struct types.

Updates #35.