ir/constant: output int const in hex notation if more readable
To improve readability of integer constants in LLVM IR output, a notion
of "low entropy" is used to determine whether to output integer
constants in decimal or hexadecimal notation.

The main idea is to check how many unique digits are required for each
representation and use the notation (hex or decimal) that requires the
least.

To illustrate this idea, here are a few examples from before and after
this commit.

Before: 2147483648
After:  u0x80000000

Before: 9218868437227405312
After:  u0x7FF0000000000000

For now, we have a threshold at 1000, so integer constants below 1000
are always represented in decimal notation. This includes negative
integer values.

Fixes #125.