ir/constant: refine notion of entropy to unique_digits/total_digits
Prior to this commit, the length of an integer constant was not
considered when computing its entropy. Thus long hexadecimal literals
got an unfair advantage as each of their digits contained more
information than their decimal counterpart.

The new definition of entropy is always in range (0.0, 1.0]. That is,
it is always above 0.0 and below or equal to 1.0.

Also, refine lower threshold to 0x1000 (instead of 1000).

If decimal notation require 0.3 or more entropy than hex notation, then
we output in hex.