ir/constant: output int const in hex notation if more readable (#126)
* ir/constant: add test case for hex vs. decimal integer notation

Updates #125.

* ir/constant: output int const in hex notation if more readable

To improve readability of integer constants in LLVM IR output, a notion
of "low entropy" is used to determine whether to output integer
constants in decimal or hexadecimal notation.

The main idea is to check how many unique digits are required for each
representation and use the notation (hex or decimal) that requires the
least.

To illustrate this idea, here are a few examples from before and after
this commit.

Before: 2147483648
After:  u0x80000000

Before: 9218868437227405312
After:  u0x7FF0000000000000

For now, we have a threshold at 1000, so integer constants below 1000
are always represented in decimal notation. This includes negative
integer values.

Fixes #125.

* ir/constant: refine notion of entropy to unique_digits/total_digits

Prior to this commit, the length of an integer constant was not
considered when computing its entropy. Thus long hexadecimal literals
got an unfair advantage as each of their digits contained more
information than their decimal counterpart.

The new definition of entropy is always in range (0.0, 1.0]. That is,
it is always above 0.0 and below or equal to 1.0.

Also, refine lower threshold to 0x1000 (instead of 1000).

If decimal notation require 0.3 or more entropy than hex notation, then
we output in hex.

* ir/constant: cap length by base when computing entropy

Without this fix, longer integer constants would get lower entropy even
though they may contain a lot of different digits.

For instance, the integer constant 1234567890 would get 1.0 as entropy,
while 12345678901234567890 would get 0.5 (since it contains every
digit, 10 and has a length of 20).

With this fix 12345678901234567890 has entropy 1.0, as should be, since
length is now capped by the base (10 for decimal).

* ir/constnat: add max hex entropy check

	Maximum allowed entropy of hexadecimal notation to output x in hexadecimal
	notation.

	This is useful as some hex values, while lower entropy than their decimal
	counter-part do not improve readability.

	For instance, the decimal entropy of 7240739780546808700 is 9/10 = 0.9 and
	the hexadecimal entropy of 0x647C4677A2884B7C is 8/16 = 0.5. As such the
	entropy difference is 0.9-0.5 = 0.4, but the hexadecimal notation does not
	improve readability. Thus we add an upper bound on the hexadecimal entropy,
	and if the entropy is above this bound, output in decimal notation
	instead.

* ir/constant: fine-tune calcMaxHexEntropy and set minEntropyDiff = 0.2

This seems to be a good balance between finding opportunities to print
values such as 0x80000 in hex while keeping values such as
0x6160ff9e9f006161 in decimal form, as they don't improve readability
enough to justify hex notation.

Now printed as hex:
	(unique hex digits: 2) hex: (length=5) 0x80000 | dec: 524288 (length=6) | hex entropy: 0.4 (max hex entropy: 0.4) | dec entropy: 0.6666666666666666 (unique digits: 4) |    ---> print as hex: 0x80000

Now printed as dec:
	(unique hex digits: 6) hex: (length=16) 0x6160ff9e9f006161 | dec: 7016889276180750689 (length=19) | hex entropy: 0.375 (max hex entropy: 0.375) | dec entropy: 0.8 (unique dec digits: 8) |    ---> print as dec: 7016889276180750689

* ir/constnat: update comment of calcMaxHexEntropy

* testdata: update to latest commit