asm: add test cases for LLVM 10.0
This includes test cases for the use of nofree as parameter attribute
and the new calling conventions of LLVM 10.0.

Updates #132.

An additional change, is to name unnamed basic blocks (if not the
entry basic block). This change was introduced in LLVM 9.0 I think.

Essentially, for any block with a local ID that is not the entry block,
we use an explicit label (e.g. `2:`) instead of just adding a comment
(e.g. `; <label>:2`. This will require an update to the llir/llvm/ir
basic block output function.