ir/constant: remove Simple from constant.Expression interface
As discussed in https://github.com/llir/llvm/issues/113#issuecomment-850797736,
the functionality of constant folding was never properly implemented
so it is very unlikely that users have come to rely on it. For this
reason, it will be removed from llir/llvm/ir/constant in its entirety.
The functionality of constant folding has been moved to irutil where
it will be developed further.

Fixes #113.