ir: use value.Value for address operand of NewIndirectBr
The address operand of indirectbr instructions should be derived
from blockaddress constants.

Note: at least in LLVM 13.0, there are no validation nor restrictions
in place to ensure that the address provided is indeed a blockaddress,
nor that it is present in the list of valid block address targets of
the indirectbr instruction.

ref: https://github.com/llir/llvm/issues/217#issuecomment-1046865923

Fixes #217.